Code/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
pyyaml>=6.0
einops==0.7.0
opencv-python==4.8.1.78
tqdm==4.66.3
pyiqa==0.1.8
matplotlib>=3.5.0
wandb>=0.12.0
scikit-image>=0.19.0
pillow>=9.0.0
